started implementation of view model
