+++ app/build.gradle
@@ -32,6 +32,7 @@ dependencies {
     implementation 'androidx.core:core-ktx:1.2.0'
     implementation 'androidx.constraintlayout:constraintlayout:1.1.3'
     implementation 'com.github.rtoshiro.mflibrary:mflibrary:1.0.0'
+    implementation "androidx.lifecycle:lifecycle-extensions:2.2.0"
     testImplementation 'junit:junit:4.12'
     androidTestImplementation 'androidx.test.ext:junit:1.1.1'
     androidTestImplementation 'androidx.test.espresso:espresso-core:3.2.0'
